Allow amending a commit when rebasing
This occurs when choosing to "edit" a commit during a rebase. Looking at git's internal shell scripts, checking if "rebase-merge/amend" exists in the .git dir is the prefered way to determine if we are in the middle of rebase and can a amend a commit.

Although we still do not allow spliting of commits while rebasing yet.